Get rid of the FiltersHelper service
